Refs #317 @@getVocabulary now uses ?resultsonly for pat-autosuggest
@gyst suggested doing it this way to keep it simple.
