+++ src/ploneintranet/theme/content/templates/document_view.pt
@@ -55,7 +55,7 @@
                 <fieldset class="pat-collapsible closed meta-extra" data-pat-collapsible="trigger: .meta-data-toggle" id="meta-extra">
                   <fieldset class="bar">
                     <input type="text" tal:attributes="disabled read_only" class="pat-autosuggest" name="subjects" placeholder="Tags" value="${python:','.join(context.subject)}"
-                           data-pat-autosuggest="ajax-data-type: json; ajax-search-index: text; ajax-search-index: results; ajax-url:${context/absolute_url}/@@getVocabulary?name=ploneintranet.network.vocabularies.Keywords"/>
+                           data-pat-autosuggest="ajax-data-type: json; ajax-search-index: text; ajax-search-index: results; ajax-url:${context/absolute_url}/@@getVocabulary?resultsonly=1&amp;name=ploneintranet.network.vocabularies.Keywords"/>
                   </fieldset>
                   <fieldset class="bar description">
                     <textarea name="description" tal:attributes="disabled read_only" rows="8" title="Description" placeholder="Description">${context/description}</textarea>
