+++ src/ploneintranet/workspace/browser/configure.zcml
@@ -198,12 +198,4 @@
 
    <include package=".tiles" />
 
-   <browser:page
-       name="getVocabulary"
-       for="*"
-       class=".vocabulary.WorkspaceVocabularyView"
-       permission="zope2.View"
-       layer="..interfaces.IPloneintranetWorkspaceLayer"
-       />
-
 </configure>
